+++ tools/etl/tg-jdbc-driver/tg-jdbc-driver/src/test/java/com/tigergraph/jdbc/log/TGLoggerFactoryTest.java
@@ -8,17 +8,17 @@ import static org.junit.Assert.*;
 public class TGLoggerFactoryTest {
 
   @Test
-  public void shouldGetJULByDefault() {
-    TGLoggerFactory.initializeLogger(1);
+  public void shouldGetJULByDefault() throws Exception {
+    TGLoggerFactory.initializeLogger(1, null);
     Logger logger = TGLoggerFactory.getLogger(TGLoggerFactoryTest.class);
     assertTrue(logger instanceof JULAdapter);
     assertEquals("JUL_DEFAULT", TGLoggerFactory.getLoggerType());
   }
 
   @Test
-  public void shouldGetJULWithNonSLF4JProperty() {
+  public void shouldGetJULWithNonSLF4JProperty() throws Exception {
     System.setProperty("com.tigergraph.jdbc.loggerImpl", "arbitrary_impl");
-    TGLoggerFactory.initializeLogger(1);
+    TGLoggerFactory.initializeLogger(1, null);
     Logger logger = TGLoggerFactory.getLogger(TGLoggerFactoryTest.class);
     assertTrue(logger instanceof JULAdapter);
     assertEquals("JUL_DEFAULT", TGLoggerFactory.getLoggerType());
@@ -26,9 +26,9 @@ public class TGLoggerFactoryTest {
   }
 
   @Test
-  public void shouldGetJULWithConfig() {
+  public void shouldGetJULWithConfig() throws Exception {
     System.setProperty("java.util.logging.config.file", "path_to_config_file");
-    TGLoggerFactory.initializeLogger(1);
+    TGLoggerFactory.initializeLogger(1, null);
     Logger logger = TGLoggerFactory.getLogger(TGLoggerFactoryTest.class);
     assertTrue(logger instanceof JULAdapter);
     assertEquals("JUL_WITH_CONFIG", TGLoggerFactory.getLoggerType());
@@ -36,9 +36,9 @@ public class TGLoggerFactoryTest {
   }
 
   @Test
-  public void shouldGetSLF4JLoggerWithSLF4JProperty() {
+  public void shouldGetSLF4JLoggerWithSLF4JProperty() throws Exception {
     System.setProperty("com.tigergraph.jdbc.loggerImpl", "SLF4J");
-    TGLoggerFactory.initializeLogger(1);
+    TGLoggerFactory.initializeLogger(1, null);
     Logger logger = TGLoggerFactory.getLogger(TGLoggerFactoryTest.class);
     assertFalse(logger instanceof JULAdapter);
     assertEquals("SLF4J", TGLoggerFactory.getLoggerType());
@@ -46,9 +46,9 @@ public class TGLoggerFactoryTest {
   }
 
   @Test
-  public void shouldGetSLF4JLoggerWithSLF4JPropertyIgnoreCase() {
+  public void shouldGetSLF4JLoggerWithSLF4JPropertyIgnoreCase() throws Exception {
     System.setProperty("com.tigergraph.jdbc.loggerImpl", "sLf4J");
-    TGLoggerFactory.initializeLogger(1);
+    TGLoggerFactory.initializeLogger(1, null);
     Logger logger = TGLoggerFactory.getLogger(TGLoggerFactoryTest.class);
     assertFalse(logger instanceof JULAdapter);
     assertEquals("SLF4J", TGLoggerFactory.getLoggerType());